Refine navigation module structure
