+++ package/lib/smooth_sheets.dart
@@ -18,6 +18,5 @@ export 'src/foundation/sheet_viewport.dart';
 export 'src/foundation/theme.dart';
 export 'src/modal/cupertino.dart';
 export 'src/modal/modal_sheet.dart';
-export 'src/navigation/navigation_routes.dart';
-export 'src/navigation/navigation_sheet.dart';
+export 'src/navigation/navigation.dart';
 export 'src/scrollable/scrollable.dart';
